funcionalidad del contador de productos y validaciones para no agregar multiples veces el mismo producto
