+++ my-app/src/api/products.json
@@ -4,73 +4,85 @@
             "id": 1,
             "product": "café americano",
             "price": 5,
-            "category": "desayuno"
+            "category": "desayuno",
+            "stock": ""
         },
         {
             "id": 2,
             "product": "café con leche",
             "price": 7,
-            "category": "desayuno"
+            "category": "desayuno",
+            "stock": ""
         },
         {
             "id": 3,
             "product": "sandwich de jamón y queso ",
             "price": 10,
-            "category": "desayuno"
+            "category": "desayuno",
+            "stock": ""
         },
         {
             "id": 4,
             "product": "jugo de frutas natural",
             "price": 7,
-            "category": "desayuno"
+            "category": "desayuno",
+            "stock": ""
         },
         {
             "id": 5,
             "product": "Hamburguesa simple",
             "price": 10,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 6,
             "product": "Hamburguesa doble",
             "price": 15,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 7,
             "product": "Papas fritas",
             "price": 5,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 8,
             "product": "Aros de cebolla",
             "price": 5,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 9,
             "product": "Agua 500 ML",
             "price": 5,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 10,
             "product": "Agua 750 ML",
             "price": 7,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 11,
             "product": "Bebida/Gaseosa 500 ML",
             "price": 7,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         },
         {
             "id": 12,
             "product": "Bebida/Gaseosa 750 ML",
             "price": 10,
-            "category": "almuerzo"
+            "category": "almuerzo",
+            "stock": ""
         }
     ]
 }
